backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.9
pydantic==2.6.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1
python-multipart==0.0.9
python-dotenv==1.0.1
psycopg2-binary==2.9.9
bcrypt==4.0.1
starlette==0.26.1
httpx==0.24.0
requests==2.32.3 